Regular commit
I fixed bags and corrected your comments
